example/Completion/app/requirements.txt
fastapi
uvicorn
python-jose
PyJWT
pydantic
